requirements.txt
openai
langchain
chromadb
spacy
uvicorn
fastapi
pydantic
pandas
requests
numpy
python-dotenv
streamlit
beautifulsoup4
pinecone-client
pymongo